Fixed another the stupidest mistake

+++ RCOM_T6_G12/parseNbuild.h
@@ -3,7 +3,7 @@
 #include "frameStructs.h"
 
 
-unsigned char *buildControlFrame(char ctrlField, unsigned fileSize, char* fileName, unsigned int L1, unsigned int L2, unsigned int frameSize);
+unsigned char *buildControlFrame(char ctrlField, unsigned long fileSize, char* fileName, unsigned int L1, unsigned int L2, unsigned int frameSize);
 
 controlFrame parseControlFrame(unsigned char *rawBytes, int size);
